create a component for managing specfic parts of the pallete
